init and nextcloud compose changes

+++ init.sh
@@ -4,11 +4,43 @@ RED="tput setaf 1"
 GREEN="tput setaf 2"
 RESET="tput sgr0"
 
-#This file is intented to configure a system from the begining
 echo -e "\nStart System Initial Configuration\n"
-sudo apt-get update
-sudo apt-get install git
-ssh-keyscan github.com >> ~/.ssh/known_hosts
+
+#This file is intented to configure a system from the begining
+read -r -p "Do you want to update the system and install recommended programs? [Y/n] " input
+case $input in
+    [yY] )
+        sudo apt-get update
+        #docker
+        sudo apt-get remove docker docker-engine docker.io
+        sudo apt-get install -y docker.io
+        sudo systemctl start docker
+        sudo systemctl enable docker
+        #git
+        sudo apt-get install git
+        if ! grep -q github.com "~/.ssh/known_hosts"; then
+            ssh-keyscan github.com >> ~/.ssh/known_hosts
+            echo "github.com was added to known hosts of ssh"
+        fi
+        ;;
+    * )
+        ;;
+esac
+
+#clone repository with setup configurations
+read -r -p "Do you want to clone linux-docker-setup? [Y/n] " input
+case $input in
+    [yY] )
+        cd ~
+        git clone git@github.com:joels-rep/linux-docker-setup.git
+        mv linux-docker-setup docker
+        cp ./docker/.vimrc .
+        cp ./docker/.bashrc .
+        sudo cp ./docker/.vimrc /root/
+        ;;
+    * )
+        ;;
+esac
 
 #echo -e "1- Download .vimrc configuration file\n"
 #FILE=~/.vimrc
@@ -32,13 +64,14 @@ case $input in
             echo "{$GREEN}PUBLIC KEY MUST BE COPIED INTO YOUR GITHUB:{$RESET}\n"
             cat ~/.ssh/id_rsa.pub >> $FILE
             cat ~/.ssh/id_rsa.pub
+            sudo systemctl restart ssh
             echo "$FILE was created successfully."
             read -p "Press Enter to continue" enter
         else
             echo "$FILE is already present in the home folder."
         fi
         ;;
-    [nN])
+    * )
         ;;
 esac
 
@@ -46,10 +79,10 @@ esac
 read -r -p "Do you want to install docker and configure containers? [Y/n] " input
 case $input in
     [yY])
-        sudo apt-get install docker-compose
-        cd ~
-        git clone git@github.com:joels-rep/linux-docker-setup.git
-        mv linux-docker-setup docker
+        #Install docker-compose
+        sudo curl -L "https://github.com/docker/compose/releases/download/v2.3.0/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
+        sudo chmod +x /usr/local/bin/docker-compose
+        #sudo ln -s /usr/local/bin/docker-compose /usr/bin/docker-compose
 
         #----------CVCode----------
         read -r -p "Do you want configure CSCode Container? [Y/n] " input
@@ -59,7 +92,7 @@ case $input in
                 cd ./docker/vscode
                 docker-compose up -d --build
                 ;;
-            [nN])
+            * )
                 ;;
         esac
         #----------NextCloud----------
@@ -68,11 +101,12 @@ case $input in
             [yY])
                 cd ~
                 cd ./docker/nextcloud
+                sudo blkid
                 read -r -p "Write the partion UUID to mount: " partition
                 sudo echo "PARTUUID=$partition  /mnt/nextcloud_disk/nextcloud ntfs defaults,noatime,uid=1000,gid=1000,dmask=007 0 0" >> /etc/fstab
                 docker-compose up -d
                 ;;
-            [nN])
+            * )
                 ;;
         esac
         #----------Portainer----------
@@ -83,11 +117,11 @@ case $input in
                 cd ./docker/portainer
                 docker-compose up -d
                 ;;
-            [nN])
+            * )
                 ;;
         esac
         ;;
-      [nN])
+      * )
           ;;
 esac
